+++ src/lib.rs
@@ -162,23 +162,23 @@ fn parse_xs_duration(s: &str) -> Result<Duration, DashMpdError> {
             }
             if let Some(s) = m.name("hours") {
                 let hours = s.as_str().parse::<u64>().unwrap();
-                secs += hours as u64 * 60 * 60;
+                secs += hours * 60 * 60;
             }
             if let Some(s) = m.name("days") {
                 let days = s.as_str().parse::<u64>().unwrap();
-                secs += days as u64 * 60 * 60 * 24;
+                secs += days * 60 * 60 * 24;
             }
             if let Some(s) = m.name("weeks") {
                 let weeks = s.as_str().parse::<u64>().unwrap();
-                secs += weeks as u64 * 60 * 60 * 24 * 7;
+                secs += weeks * 60 * 60 * 24 * 7;
             }
             if let Some(s) = m.name("months") {
                 let months = s.as_str().parse::<u64>().unwrap();
-                secs += months as u64 * 60 * 60 * 24 * 30;
+                secs += months * 60 * 60 * 24 * 30;
             }
             if let Some(s) = m.name("years") {
                 let years = s.as_str().parse::<u64>().unwrap();
-                secs += years as u64 * 60 * 60 * 24 * 365;
+                secs += years * 60 * 60 * 24 * 365;
             }
             if let Some(s) = m.name("sign") {
                 if s.as_str() == "-" {
